Bump release number
# New features:

## Output images to blob form
Now you can output a blob form of the image itself, as opposed to saving it on disk somewhere. This is useful if you want to upload the images to, say, imgur, as you can pipe it right into a multipart form upload!

## Open flag on `meme`

This was one that was requested by a friend. Now you can run the `meme` command with the `-O` (capital o) flag, and I'll try to open it in your system's default image viewer. Confirmed working on OS X!